Changing over to Integer IDs Some schema changes Some converter fixed for Locality Added some new Combo boxes.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@2590 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/dbsupport/DataProviderSessionIFace.java
@@ -44,9 +44,9 @@ public interface DataProviderSessionIFace
     
     //public List getDataList(RecordSetIFace recordSet);
     
-    public <T> T load(Class<T> clsObj, Long id);
+    public <T> T load(Class<T> clsObj, Integer id);
     
-    public <T> T get(Class<T> clsObj, Long id);
+    public <T> T get(Class<T> clsObj, Integer id);
 
     public <T> T getData(Class<T> clsObject, String fieldName, Object value, DataProviderSessionIFace.CompareType compareType);
