+++ src/edu/ku/brc/specify/datamodel/TreeDefIface.java
@@ -52,8 +52,8 @@ public interface TreeDefIface<N extends Treeable<N,D,I>,
 
     public void initialize();
 	
-	public Long getTreeDefId();
-	public void setTreeDefId(Long id);
+	public Integer getTreeDefId();
+	public void setTreeDefId(Integer id);
 	
 	public String getRemarks();
 	public void setRemarks(String remarks);
